Added unit test for ^^ageListViewState

+++ app/src/main/java/com/caglacetin/lorempicsum/data/response/ImageData.kt
@@ -8,8 +8,8 @@ import kotlinx.android.parcel.Parcelize
 data class ImageData(
   val id: String = "",
   val author: String = "",
-  val width: Int = 0,
-  val height: Int = 0,
+  val width: Long = 0,
+  val height: Long = 0,
   val url: String = "",
   @Json(name = "download_url")
   val downloadUrl: String = ""
